einige Objekte und Attribute umbenannt

+++ 2_Quellcode/Server/src/Server/ServThread.java
@@ -17,18 +17,22 @@ public class ServThread implements Runnable{
 	 protected String serverText   = null;
 	 private BufferedReader reader;
 	 private PrintWriter printWriter;
-	 private dateiHandler fehler;
-	 private  dataBaseHandler dbh ;
+	 private DateiHandler fehler;
+	 private  DataBaseHandler dbh ;
+	 private boolean adminLoggedIn;
 	 
 
-	    public ServThread(java.net.Socket clientSocket, dataBaseHandler dbh) {
+	    public ServThread(java.net.Socket clientSocket, DataBaseHandler dbh, boolean admin) {
 	    	this.dbh = dbh;
 	        this.clientsocket = clientSocket;
+	        this.adminLoggedIn = admin;
+	        fehler = new DateiHandler("THREADerrLog.txt");
 	        try {
 				reader = new BufferedReader(new InputStreamReader(clientsocket.getInputStream()));
 				printWriter = new PrintWriter(new OutputStreamWriter(clientsocket.getOutputStream()));
-				fehler = new dateiHandler("THREADerrLog.txt");
+				
 			} catch (IOException e) {
+				
 				fehler.openDatei(true);
 				fehler.writeErr(e.getMessage()+ "\n");
 				fehler.close();
@@ -67,7 +71,7 @@ public class ServThread implements Runnable{
 			newEvent2DB(cmd); // /INSERT/Bzeichner/tt-mm-jjjj/state
 		}
 		else if(Parser.parseCMD(cmd) == 3) {
-			modifiyEvent2DB(cmd); // /MODIFY/id/Bzeichner/tt-mm-jjjj/state
+			modifiyEvent2DB(cmd); // /MODIFY/id/Bzeichner/tt-mm-jjjj/state/prot
 		}
 		else if(Parser.parseCMD(cmd) == 4) {
 			deleteEntry2DB(cmd); // /DELETE/id
@@ -84,13 +88,12 @@ public class ServThread implements Runnable{
 	
 	public void doUpdate2DB() {
 		int i = 1;
-		System.out.println("UPDATE");
 		String buffer[] = new String[4];
 		try {
 			buffer = dbh.getEntry(i);
 			while(!(buffer[0] == null)) {
 				i++;
-				putStr("," +buffer[0]+ "," + buffer[1] + "," + buffer[2]+ "," + "\n");
+				putStr("," +buffer[0]+ "," + buffer[1] + "," + buffer[2]+ "," + buffer[3] + "," + "\n");
 				buffer = dbh.getEntry(i);
 			}
 			
@@ -108,8 +111,14 @@ public class ServThread implements Runnable{
 		String bezeichner = attr[1];
 		String date = attr[2];
 		String state = attr[3];
+		String prot = attr[4];
 		int id = dbh.getNumRows()+1;
-		dbh.insertEntry(id, bezeichner, date, state);
+		if(!this.adminLoggedIn) {
+			dbh.insertEntry(id, bezeichner, date, state, "false");
+		}
+		else {
+			dbh.insertEntry(id, bezeichner, date, state, prot);
+		}
 	}
 	
 	public void modifiyEvent2DB(String cmd) {
@@ -119,13 +128,26 @@ public class ServThread implements Runnable{
 		String bezeichner = attr[1];
 		String date = attr[2];
 		String state = attr[3];
-		dbh.modifyEntry(id, bezeichner, date, state);
+		String prot = attr[4];
+		if(!this.adminLoggedIn) {
+		String buffPriv = Boolean.toString(dbh.getPriv(id));
+		dbh.modifyEntry(id, bezeichner, date, state, buffPriv);
+		}
+		else {
+			dbh.modifyEntry(id, bezeichner, date, state, prot);
+		}
 	}
 	
 	public void deleteEntry2DB(String cmd) {
 		int id = Parser.getID(cmd);
+		if(dbh.getPriv(id) && this.adminLoggedIn) {
 		dbh.deleteEntry(id);
 		dbh.udpdateIDs();
+		}
+		else if(!dbh.getPriv(id)){
+			dbh.deleteEntry(id);
+			dbh.udpdateIDs();
+		}
 	}
 	
 }
